Task_3/requirements.txt
flask==2.3.3
kafka-python==2.0.2
requests==2.31.0